Add tooltips to character portrait, add tooltip for block.
There is some weird behavior (likely a bug) with tooltip not growing
enough to fit text when set to "Shrink vertically", so I moved it inside
a PanelContainer that replaces the image.

As of now, tooltips for a given skill live in card_ui and the tooltips
in the character_portrait ui, it would be nice to put them in a single
place.

Fixes #34.