Add a new CardEffect implementation that allows for more flexibility.
Implement New Plan with it, a card that allows to discard your hand
and draw back as many cards as you had.

CardEffectNew works as a proof of concept, but it needs to get support
for all the previous fields, etc that the previous cards need to
support.

For now both modes work (but shouldn't be mixed in the same card just
in case).

Beginning of #52.