Add StatsManager singleton.
It supports multiple levels of tracking (overall, run, stae, turn, etc).

This allows to easily keep track of stats from anywhere without having
to pass stats objects around and to get those stats to e.g. be used by
cards.

All the stats in the enum shoudl be correctly tracked as of now.

Work for #57.